requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.0.3
django-cors-headers==4.3.1
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
pillow==10.2.0
proglog==0.1.10
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
